requirements.txt
# DataSentinel AI - Requirements
# ================================
# Généré depuis pyproject.toml pour compatibilité pip install -r

# Core
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LangChain ecosystem
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.10

# Vector store
chromadb>=0.4.0

# Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
pyarrow>=14.0.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Observability
prometheus-fastapi-instrumentator>=0.11.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Session persistence
redis>=5.0.0

# Rate limiting
slowapi>=0.1.9

# PDF generation
reportlab>=4.0.0

# Excel generation
openpyxl>=3.1.0

# UI
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
structlog>=23.1.0
httpx>=0.25.0
requests>=2.31.0

# Dev dependencies (uncomment if needed)
 pytest>=7.4.0
 pytest-asyncio>=0.21.0
 pytest-cov>=4.1.0
 fakeredis>=2.20.0
 black>=23.0.0
 ruff>=0.1.0
 mypy>=1.5.0
